requirements.txt
streamlit
azure-ai-projects>=2.0.0b1
azure-identity
python-dotenv